+++ src/core/data/modules/base.ts
@@ -286,6 +286,8 @@ export default abstract class Provider extends ParamsProvider implements IProvid
 
 		requestCache[this.cacheId] = Object.createDict();
 		this.async.terminateWorker({group: 'extraProvidersCache'});
+
+		this.emitter.emit('dropCache', recursive);
 	}
 
 	/** @inheritDoc */
@@ -309,6 +311,8 @@ export default abstract class Provider extends ParamsProvider implements IProvid
 		Object.keys(this.params).forEach((key, _, data) => {
 			delete data[key];
 		});
+
+		this.emitter.emit('destroy');
 	}
 
 	/** @inheritDoc */