+++ src/core/request/engines/provider/index.ts
@@ -35,6 +35,8 @@ import type { RequestEngine, RequestOptions } from 'core/request/interface';
 import { availableParams, deniedProviderParams } from 'core/request/engines/provider/const';
 import type { AvailableOptions, MethodsMapping, Meta } from 'core/request/engines/provider/interface';
 
+export * from 'core/request/engines/composition';
+
 export * from 'core/request/engines/provider/const';
 export * from 'core/request/engines/provider/interface';
 
@@ -210,6 +212,7 @@ export default function createProviderEngine(
 				important: providerResponse.important,
 
 				okStatuses: providerResponse.okStatuses,
+				noContentStatuses: providerResponse.noContentStatuses,
 				status: providerResponse.status,
 				statusText: providerResponse.statusText,